llama_cpp server: app is now importable, still runnable as a module
